Combined the Files and Workspace Pane
* Files pane is now deprecated in favour of Workspace pane
* Renamed IDEWorkspace to WorkspacePane
